+++ interfaces/http/handler/account_handler.go
@@ -1,0 +1,53 @@
+package handler
+
+import (
+	"github.com/gin-gonic/gin"
+	"github.com/kiosanim/pismo-code-assessment/application/account/dto"
+	"github.com/kiosanim/pismo-code-assessment/internal/domains/account"
+	"net/http"
+	"strconv"
+)
+
+type AccountHandler struct {
+	service account.Service
+}
+
+func NewAccountHandler(service account.Service) *AccountHandler {
+	return &AccountHandler{service: service}
+}
+
+func (h *AccountHandler) CreateAccount(c *gin.Context) {
+	var req dto.CreateAccountRequest
+	if err := c.ShouldBindJSON(&req); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	res, err := h.service.Create(c.Request.Context(), req)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusCreated, res)
+}
+
+func (h *AccountHandler) GetAccountByID(c *gin.Context) {
+
+	accountId := c.Param("account_id")
+	accountId64, err := strconv.ParseInt(accountId, 10, 64)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	res, err := h.service.FindByID(c.Request.Context(), dto.FindAccountByIdRequest{AccountID: accountId64})
+	if res == nil && err != nil {
+		c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
+		return
+	} else if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusCreated, res)
+}