+++ internal/domains/account/repository.go
@@ -9,5 +9,5 @@ var AccountRepositoryInvalidParametersError = errors.New("account repository inv
 
 type AccountRepository interface {
 	FindByID(ctx context.Context, accountID int64) (*Account, error)
-	Save(ctx context.Context, account *Account) (*Account, error)
+	Save(ctx context.Context, newAccount *Account) (*Account, error)
 }