+++ internal/domains/transaction/repository.go
@@ -8,5 +8,5 @@ import (
 var TransactionRepositoryInvalidParametersError = errors.New("transaction repository invalid parameters")
 
 type TransactionRepository interface {
-	Save(ctx context.Context, transaction *Transaction) (*Transaction, error)
+	Save(ctx context.Context, newTransaction *Transaction) (*Transaction, error)
 }