+++ CHANGELOG.md
@@ -4,5 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] 2025-11-13
+## [0.3.0] 2025-11-13
+- Adding Account Handler
+- Adding Transaction Handler
+- Adding Router
+- Adding Migration Tool
+
+## [0.2.0] 2026-01-24
+- Adding Account Service
+- Adding Transaction Service
+- Adding Account Repository
+- Adding Transaction Repository
+
+## [0.1.0] 2026-01-23
 - Initial Version